refactor: rename the hook to create new steps
