+++ packages/twenty-front/src/modules/workflow/hooks/useRightDrawerWorkflowSelectActionCreateStep.tsx
@@ -1,15 +1,20 @@
-import { useCreateNode } from '@/workflow/hooks/useCreateNode';
+import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { useUpdateOneRecord } from '@/object-record/hooks/useUpdateOneRecord';
 import { showPageWorkflowDiagramTriggerNodeSelectionState } from '@/workflow/states/showPageWorkflowDiagramTriggerNodeSelectionState';
 import { workflowCreateStepFromParentStepIdState } from '@/workflow/states/workflowCreateStepFromParentStepIdState';
 import {
   Workflow,
   WorkflowStep,
   WorkflowStepType,
+  WorkflowVersion,
 } from '@/workflow/types/Workflow';
+import { getWorkflowLastVersion } from '@/workflow/utils/getWorkflowLastVersion';
+import { insertStep } from '@/workflow/utils/insertStep';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { isDefined } from 'twenty-ui';
 import { v4 } from 'uuid';
 
-export const useRightDrawerWorkflowSelectAction = ({
+export const useRightDrawerWorkflowSelectActionCreateStep = ({
   workflow,
 }: {
   workflow: Workflow;
@@ -22,7 +27,36 @@ export const useRightDrawerWorkflowSelectAction = ({
     showPageWorkflowDiagramTriggerNodeSelectionState,
   );
 
-  const { createNode } = useCreateNode({ workflow });
+  const { updateOneRecord: updateOneWorkflowVersion } =
+    useUpdateOneRecord<WorkflowVersion>({
+      objectNameSingular: CoreObjectNameSingular.WorkflowVersion,
+    });
+
+  const insertNodeAndSave = ({
+    parentNodeId,
+    nodeToAdd,
+  }: {
+    parentNodeId: string;
+    nodeToAdd: WorkflowStep;
+  }) => {
+    const lastVersion = getWorkflowLastVersion(workflow);
+    if (!isDefined(lastVersion)) {
+      throw new Error(
+        "Can't add a node when no version exists yet. Create a first workflow version before trying to add a node.",
+      );
+    }
+
+    return updateOneWorkflowVersion({
+      idToUpdate: lastVersion.id,
+      updateOneRecordInput: {
+        steps: insertStep({
+          steps: lastVersion.steps,
+          parentStepId: parentNodeId,
+          stepToAdd: nodeToAdd,
+        }),
+      },
+    });
+  };
 
   const getStepDefaultConfiguration = (
     type: WorkflowStepType,
@@ -55,7 +89,7 @@ export const useRightDrawerWorkflowSelectAction = ({
     }
   };
 
-  const handleActionClick = async (newStepType: WorkflowStepType) => {
+  const createStep = async (newStepType: WorkflowStepType) => {
     if (workflowCreateStepFromParentStepId === undefined) {
       throw new Error('Select a step to create a new step from first.');
     }
@@ -65,14 +99,14 @@ export const useRightDrawerWorkflowSelectAction = ({
     /**
      * FIXME: For now, the data of the node to create are mostly static.
      */
-    await createNode({
+    await insertNodeAndSave({
       parentNodeId: workflowCreateStepFromParentStepId,
       nodeToAdd: newStep,
     });
 
     /**
      * After the step has been created, select it.
-     * As the `createNode` function mutates the cached workflow before resolving,
+     * As the `insertNodeAndSave` function mutates the cached workflow before resolving,
      * we are sure that the new node will have been created at this stage.
      *
      * Selecting the node will cause a right drawer to open in order to edit the step.
@@ -81,6 +115,6 @@ export const useRightDrawerWorkflowSelectAction = ({
   };
 
   return {
-    handleActionClick,
+    createStep,
   };
 };
